Add files via upload
At IWB Technology, we are committed to promoting sustainability and environmental responsibility by providing comprehensive electronic waste (e-waste) recycling solutions. We believe in the power of technology to create a greener future, and our services are designed to help businesses, individuals, and organizations responsibly recycle outdated or broken electronic devices.

As a leading e-waste recycling company, we specialize in recovering valuable materials from electronic waste, while ensuring the safe disposal of hazardous components. Our mission is to reduce the harmful impact of electronic waste on the environment, conserve natural resources, and contribute to a circular economy.
At IWB Technology, we understand the importance of raising awareness about the environmental impact of electronic waste. That’s why we actively engage in community education and outreach programs, such as:

Hosting e-waste collection events in local communities
Organizing workshops and seminars to educate the public on the importance of recycling
Promoting eco-friendly practices in schools, businesses, and households
We are dedicated to making e-waste recycling accessible and beneficial for everyone.